Refactor Home component and introduce useTodoData hook
- Replaced local state management in Home component with useTodoData hook for better separation of concerns.
- Added ITEMS_PER_PAGE constant for pagination consistency.
- Updated modal handling to use handleModalClose from useTodoData.
- Cleaned up useHome hook by removing unused state and functions, streamlining todo data management.
