+++ app/constants/common.ts
@@ -1,0 +1,1 @@
+export const ITEMS_PER_PAGE = 30;
